backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
web3==6.15.1
python-dotenv==1.0.0
requests==2.31.0
praw==7.7.1
beautifulsoup4==4.12.3
